refactor(ui): move react components to react codebase [EE-3354]
