+++ app/kubernetes/views/configure/configureController.js
@@ -7,9 +7,9 @@ import KubernetesNamespaceHelper from 'Kubernetes/helpers/namespaceHelper';
 import { FeatureId } from '@/react/portainer/feature-flags/enums';
 
 import { getIngressControllerClassMap, updateIngressControllerClassMap } from '@/react/kubernetes/cluster/ingressClass/utils';
-import { getIsRBACEnabled } from '@/react/kubernetes/cluster/service';
 import { buildConfirmButton } from '@@/modals/utils';
 import { confirm } from '@@/modals/confirm';
+import { getIsRBACEnabled } from '@/react/kubernetes/cluster/getIsRBACEnabled';
 
 class KubernetesConfigureController {
   /* #region  CONSTRUCTOR */
