+++ app/portainer/components/BoxSelector/BoxSelectorAngular.ts
