+++ app/portainer/oauth/components/oauth-providers-selector/oauth-provider-selector.controller.js
@@ -1,4 +1,4 @@
-import { options } from './oauth-options';
+import { options } from '@/react/portainer/settings/AuthenticationView/oauth-options';
 
 export default class OAuthProviderSelectorController {
   constructor() {
