--- app/portainer/views/endpoints/kvm/index.js
@@ -1,5 +1,0 @@
-import angular from 'angular';
-
-import { KVMControlAngular } from '@/portainer/views/endpoints/kvm/KVMControl';
-
-angular.module('portainer.app').component('kvmControl', KVMControlAngular).name;
