+++ app/react/edge/edge-devices/WaitingRoomView/queries.ts
@@ -3,7 +3,7 @@ import { useMutation, useQueryClient } from 'react-query';
 import { EnvironmentId } from '@/react/portainer/environments/types';
 import axios, { parseAxiosError } from '@/portainer/services/axios';
 import { promiseSequence } from '@/portainer/helpers/promise-utils';
-import { useIntegratedLicenseInfo } from '@/portainer/license-management/use-license.service';
+import { useIntegratedLicenseInfo } from '@/react/portainer/licenses/use-license.service';
 
 export function useAssociateDeviceMutation() {
   const queryClient = useQueryClient();
