+++ app/react/portainer/environments/KvmView/KVMControl/index.ts
@@ -1,0 +1,1 @@
+export { KVMControl } from './KVMControl';
