+++ app/react/portainer/licenses/use-license.service.ts
