+++ app/react/portainer/registries/CreateView/options.tsx
