+++ app/react/portainer/settings/AuthenticationView/InternalAuth/options.tsx
