+++ app/react/portainer/settings/AuthenticationView/ldap-options.tsx
