+++ app/react/portainer/settings/AuthenticationView/oauth-options.tsx
